Final_Project/requirements.txt
absl-py==1.0.0
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
astunparse==1.6.3
attendance==1.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
Bashutils==0.0.4
bcrypt==3.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blurb==1.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cmake==3.22.4
colorama==0.4.4
cycler==0.11.0
DateTime==4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.4
django-crispy-forms==1.14.0
django-pandas==0.6.6
dlib @ https://github.com/jloh02/dlib/releases/download/v19.22/dlib-19.22.99-cp310-cp310-win_amd64.whl
dnspython==2.2.1
docopt==0.6.2
docutils==0.18.1
email-validator==1.1.3
entrypoints==0.4
executing==0.8.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0
fonttools==4.33.3
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.46.0
h5py==3.6.0
idna==3.3
image==1.5.33
ImageFilter==0.1.0
imageio==2.19.1
imagesize==1.3.0
importlib-metadata==4.11.3
imutils==0.5.4
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
Js2Py==0.71
jsonschema==4.5.1
jupyter==1.0.0
jupyter-client==7.3.1
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
layers==0.1.5
libclang==14.0.1
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mysqlclient==2.1.0
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
pipreqs==0.4.11
pipwin==0.5.2
platformdirs==2.5.2
playsound==1.2.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pygame==2.1.2
Pygments==2.12.0
pyjsparser==2.7.1
pyparsing==3.0.8
PyPrind==2.11.3
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dateutil==2.8.2
python-doc==2020.10.1
python-docs-theme==2022.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
sequential==1.0.0
six==1.16.0
sklearn==0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.35
sqlparse==0.4.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.13.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
wrapt==1.14.1
WTForms==3.0.1
yarg==0.1.9
zipp==3.8.0
zope.interface==5.4.0
